lets work

+++ src/memberInfo/SignUp.java
@@ -1,5 +1,6 @@
 package memberInfo;
 
+import java.io.File;
 import java.io.FileWriter;
 
 public class SignUp {
@@ -10,13 +11,18 @@ public class SignUp {
 	
 	public void save(Member m)  throws Exception
 	{
-		
-		String s = m.name+" ,"+m.id+" ,"+" ,"+m.password+" ,"+m.gender+" ,"+m.age+" ,"+m.address+"\n";
-			FileWriter y = new FileWriter("D:\\eclipse-workspace\\Project\\Records\\Member\\Member's User_Password.txt",true);
-			y.write(m.name+"\n"+m.password+"\n");
+		File f = new File("D:\\eclipse-workspace\\Project\\Records\\Member\\"+m.getId()+".txt");
+		if(f.exists() && !f.isDirectory()) { 
+			System.out.println("Already Signed up");
+			throw new SignUpException();
+		}
+		FileWriter y = new FileWriter("D:\\eclipse-workspace\\Project\\Records\\Member\\Member's User_Password.txt",true);
+			y.write(m.getName()+"\n"+m.getPassword()+"\n");
 			y.close();
-			FileWriter x = new FileWriter("D:\\eclipse-workspace\\Project\\Records\\Member\\"+m.id+".txt");
+			String s = m.getName()+"\n"+m.getId()+"\n"+m.getfName()+"\n"+m.getmName()+"\n"+m.getAddress()+"\n"+m.getAge()+"\n"+m.getGender()+"\n"+m.getPassword()+"\n";
+			FileWriter x = new FileWriter("D:\\eclipse-workspace\\Project\\Records\\Member\\"+m.getId()+".txt");
 			x.write(s);
+			System.out.println("File Created");
 			x.close();
 	}
